requirements.txt
pyTelegramBotAPI==4.14.0
Pillow==10.2.0
